Update specs for refactored state
