+++ specs/BasicTilesetProcessorSpec.ts
@@ -2,7 +2,7 @@
 import fs from "fs";
 
 import { SpecHelpers } from "./SpecHelpers";
-import { SpecProcessor } from "./SpecEntryProcessor";
+import { SpecEntryProcessor } from "./SpecEntryProcessor";
 
 import { Paths } from "../src/base/Paths";
 
@@ -76,48 +76,13 @@ describe("BasicTilesetProcessor on explicit input", function () {
     expect(actualContentUris).toEqual(expectedContentUris);
   });
 
-  it("processAllEntries processes all entries exactly once", async function () {
-    const tilesetProcessor = new BasicTilesetProcessor(quiet);
-    await tilesetProcessor.begin(basicInput, basicOutput, overwrite);
-    const specProcessor = new SpecProcessor();
-    await tilesetProcessor.processAllEntries(specProcessor.processEntry);
-    await tilesetProcessor.end();
-
-    // Expect ALL files to have been processed
-    // (except for 'tileset.json')
-    const expectedProcessedKeys = [
-      "README.md",
-      "sub/tileB.b3dm",
-      "tileA.b3dm",
-      "tileB.b3dm",
-      "tileC.b3dm",
-    ];
-    const actualProcessedKeys = specProcessor.processedKeys;
-    actualProcessedKeys.sort();
-    expect(actualProcessedKeys).toEqual(expectedProcessedKeys);
-
-    // Expect the names of ALL files to have been modified
-    // (except for 'tileset.json')
-    const expectedOutputFiles = [
-      "PROCESSED_README.md",
-      "PROCESSED_sub/tileB.b3dm",
-      "PROCESSED_tileA.b3dm",
-      "PROCESSED_tileB.b3dm",
-      "PROCESSED_tileC.b3dm",
-      "tileset.json",
-    ];
-    const actualOutputFiles = SpecHelpers.collectRelativeFileNames(basicOutput);
-    actualOutputFiles.sort();
-    expect(actualOutputFiles).toEqual(expectedOutputFiles);
-  });
-
   it("processTileContentEntries processes the tile content entries", async function () {
     const tilesetProcessor = new BasicTilesetProcessor(quiet);
     await tilesetProcessor.begin(basicInput, basicOutput, overwrite);
-    const specProcessor = new SpecProcessor();
+    const specEntryProcessor = new SpecEntryProcessor();
     await tilesetProcessor.processTileContentEntries(
-      specProcessor.processUri,
-      specProcessor.processEntry
+      specEntryProcessor.processUri,
+      specEntryProcessor.processEntry
     );
     await tilesetProcessor.end();
 
@@ -128,7 +93,7 @@ describe("BasicTilesetProcessor on explicit input", function () {
       "tileB.b3dm",
       "tileC.b3dm",
     ];
-    const actualProcessedKeys = specProcessor.processedKeys;
+    const actualProcessedKeys = specEntryProcessor.processedKeys;
     actualProcessedKeys.sort();
     expect(actualProcessedKeys).toEqual(expectedProcessedKeys);
 
@@ -149,10 +114,10 @@ describe("BasicTilesetProcessor on explicit input", function () {
   it("processTileContentEntries updates the content URIs", async function () {
     const tilesetProcessor = new BasicTilesetProcessor(quiet);
     await tilesetProcessor.begin(basicInput, basicOutput, overwrite);
-    const specProcessor = new SpecProcessor();
+    const specEntryProcessor = new SpecEntryProcessor();
     await tilesetProcessor.processTileContentEntries(
-      specProcessor.processUri,
-      specProcessor.processEntry
+      specEntryProcessor.processUri,
+      specEntryProcessor.processEntry
     );
     await tilesetProcessor.end();
 
@@ -175,29 +140,5 @@ describe("BasicTilesetProcessor on explicit input", function () {
       "PROCESSED_tileC.b3dm",
     ];
     expect(actualContentUris).toEqual(expectedContentUris);
-  });
-
-  it("processAllEntries only processes unprocessed entries", async function () {
-    const tilesetProcessor = new BasicTilesetProcessor(quiet);
-    await tilesetProcessor.begin(basicInput, basicOutput, overwrite);
-
-    // First, process all content entries
-    const contentsSpecProcessor = new SpecProcessor();
-    await tilesetProcessor.processTileContentEntries(
-      contentsSpecProcessor.processUri,
-      contentsSpecProcessor.processEntry
-    );
-
-    // Now, process all remaining entries
-    const specProcessor = new SpecProcessor();
-    await tilesetProcessor.processAllEntries(specProcessor.processEntry);
-    await tilesetProcessor.end();
-
-    // Expect only the non-content entries to have been processed
-    // in processAllEntries
-    const expectedProcessedKeys = ["README.md"];
-    const actualProcessedKeys = specProcessor.processedKeys;
-    actualProcessedKeys.sort();
-    expect(actualProcessedKeys).toEqual(expectedProcessedKeys);
   });
 });
