+++ specs/TilesetProcessorSpec.ts
@@ -1,16 +1,22 @@
-import { BasicTilesetProcessor } from "../src/tilesetProcessing/BasicTilesetProcessor";
+import { TilesetDataProcessor } from "../src/tilesetProcessing/TilesetDataProcessor";
 
+import { SpecEntryProcessor } from "./SpecEntryProcessor";
 import { SpecHelpers } from "./SpecHelpers";
 
 const basicInput = "./specs/data/tilesetProcessing/basicProcessing";
 const basicOutput = "./specs/data/output/tilesetProcessing/basicProcessing";
+
+const implicitInput = "./specs/data/tilesetProcessing/implicitProcessing";
+const implicitOutput =
+  "./specs/data/output/tilesetProcessing/implicitProcessing";
+
 const quiet = true;
 const overwrite = true;
 
 /**
  * Tests for the base functionality of the (abstract) TilesetProcessor
- * base class, using its only concrete implementation, namely the
- * BasicTilesetProcessor
+ * base class, using one concrete implementation, namely the
+ * TilesetDataProcessor
  */
 describe("TilesetProcessor", function () {
   afterEach(function () {
@@ -18,7 +24,7 @@ describe("TilesetProcessor", function () {
   });
 
   it("throws when trying to call 'begin' with invalid path", async function () {
-    const tilesetProcessor = new BasicTilesetProcessor(quiet);
+    const tilesetProcessor = new TilesetDataProcessor(quiet);
     await expectAsync(
       (async function () {
         await tilesetProcessor.begin(
@@ -33,7 +39,7 @@ describe("TilesetProcessor", function () {
   });
 
   it("throws when trying to call 'begin' twice", async function () {
-    const tilesetProcessor = new BasicTilesetProcessor(quiet);
+    const tilesetProcessor = new TilesetDataProcessor(quiet);
     await tilesetProcessor.begin(basicInput, basicOutput, overwrite);
     await expectAsync(
       (async function () {
@@ -45,7 +51,7 @@ describe("TilesetProcessor", function () {
   });
 
   it("throws when trying to call 'end' without 'begin'", async function () {
-    const tilesetProcessor = new BasicTilesetProcessor(quiet);
+    const tilesetProcessor = new TilesetDataProcessor(quiet);
     await expectAsync(
       (async function () {
         await tilesetProcessor.end();
@@ -56,7 +62,7 @@ describe("TilesetProcessor", function () {
   });
 
   it("throws when trying to call 'end' twice", async function () {
-    const tilesetProcessor = new BasicTilesetProcessor(quiet);
+    const tilesetProcessor = new TilesetDataProcessor(quiet);
     await tilesetProcessor.begin(basicInput, basicOutput, overwrite);
     await tilesetProcessor.end();
     await expectAsync(
@@ -69,7 +75,7 @@ describe("TilesetProcessor", function () {
   });
 
   it("performs a 'no-op' of just copying the data when when no other functions are called", async function () {
-    const tilesetProcessor = new BasicTilesetProcessor(quiet);
+    const tilesetProcessor = new TilesetDataProcessor(quiet);
     await tilesetProcessor.begin(basicInput, basicOutput, overwrite);
     await tilesetProcessor.end();
